use unittesting.json for settings
