+++ sbin/run.py
@@ -12,9 +12,6 @@ import json
 
 opts, args = getopt.getopt(sys.argv[1:], "ad", ["tests_dir=", "async", "deferred"])
 
-tests_dir = ([a for o,a in opts if o == "--tests_dir"] + ["tests"])[0]
-async = any([o in ['-a', '--async'] for o,a in opts])
-deferred = any([o in ['-d', '--deferred'] for o,a in opts])
 package = args[0] if len(args)>0 else "UnitTesting"
 
 version = int(subprocess.check_output(["subl","--version"]).decode('utf8').strip()[-4])
@@ -42,8 +39,7 @@ try:
 except:
     schedule = []
 if not any([s['package']==package for s in schedule]):
-    schedule.append({'package': package, 'tests_dir': tests_dir,
-        'async': async, 'deferred': deferred})
+    schedule.append({'package': package})
 with open(jpath, 'w') as f:
     f.write(json.dumps(schedule, ensure_ascii=False, indent=True))
